Organize devices page components
