+++ src/js/views/devices/layout/DevicesLoading.jsx
