+++ src/js/views/devices/layout/EmptyDeviceList.jsx
